# schemas/openapi.json
{
	"openapi": "3.0.0",
	"info": {
		"version": "1.0.0",
		"title": "Canine shelter API",
		"description": "An API for a canine shelter website",
		"contact": {
			"name": "Bartlomiej Wlodarski",
			"email": "wlodarsb@uni.coventry.ac.uk"
		},
		"license": {
			"name": "Apache 2.0",
			"url": "https://www.apache.org/licenses/LICENSE-2.0.html"
		}
	},
	"servers": [
		{
			"url": "https://canoe-janet-3000.codio-box.uk/api/v1",
			"description": "API backend server"
		}
	],
	"security": [
		{
			"basic": [],
			"jwt": []
		}
	],
...
	"components": {
		"responses": {
			"authenticationError": {
				"description": "User needs to be authenticated to access this route"
			},
			"notFoundError": {
				"description": "Resource does not exist",
				"content": {
					"application/json": {
						"schema": {
							"type": "object",
							"properties": {
								"message": {
									"description": "Error message",
									"type": "string",
									"example": "User does not exist"
								}
							}
						}
					}
				}
			},
			"validationError": {
				"description": "Occurs when the server rejects the form data",
				"content": {
					"application/json": {
						"schema": {
							"$ref": "./errors.json#/definitions/validationErrors"
						}
					}
				}
			}
		},
		"parameters": {
...
			"dogId": {
				"name": "id",
				"in": "path",
				"description": "Dog ID",
				"required": true,
				"schema": { "type": "integer" }
			},
			"locationId": {
				"name": "id",
				"in": "path",
				"description": "Location ID",
				"required": true,
				"schema": { "type": "integer" }
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			},
			"basic": {
				"type": "http",
				"scheme": "basic"
			}
		}
	},
	"paths": {
...
		"/breed/{id}/dogs": {
			"summary": "Dogs associated with breed",
			"get": {
				"tags": ["Breeds"],
				"summary": "Get a list of dogs associated with breed",
				"security": [],
				"parameters": [
					{
						"name": "query",
						"description": "A search query",
						"in": "query",
						"required": false,
						"schema": { "type": "string" },
						"example": "Rex"
					},
					{
						"name": "page",
						"description": "Page number to browse",
						"in": "query",
						"required": false,
						"schema": { "type": "integer" },
						"example": 1
					},
					{
						"name": "limit",
						"description": "Limit of entries per page",
						"in": "query",
						"required": false,
						"schema": { "type": "integer" },
						"example": 5
					},
					{
						"name": "order",
						"description": "Item to order the results by",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string",
							"enum": ["id", "name", "description", "age", "gender"]
						},
						"example": "age"
					},
					{
						"name": "direction",
						"description": "Direction to sort in (ascending or descending)",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string",
							"enum": ["asc", "desc"]
						},
						"example": "asc"
					}
				],
				"responses": {
					"200": {
						"description": "A list of dogs",
						"content": {
							"application/json": {
								"schema": { "$ref": "./dogs.json#/definitions/dogs" }
							}
						}
					},
					"404": {
						"$ref": "#/components/responses/notFoundError"
					}
				}
			}
		},
...
		"/dogs": {
			"summary": "Represents dogs on the website",
			"get": {
				"tags": ["Dogs"],
				"description": "Get a list of dogs from the database",
				"security": [],
				"parameters": [
					{
						"name": "query",
						"description": "A search query",
						"in": "query",
						"required": false,
						"schema": { "type": "string" },
						"example": "Rex"
					},
					{
						"name": "page",
						"description": "Page number to browse",
						"in": "query",
						"required": false,
						"schema": { "type": "integer" },
						"example": 1
					},
					{
						"name": "limit",
						"description": "Limit of entries per page",
						"in": "query",
						"required": false,
						"schema": { "type": "integer" },
						"example": 5
					},
					{
						"name": "order",
						"description": "Item to order the results by",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string",
							"enum": ["id", "name", "description", "age", "gender"]
						},
						"example": "age"
					},
					{
						"name": "direction",
						"description": "Direction to sort in (ascending or descending)",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string",
							"enum": ["asc", "desc"]
						},
						"example": "asc"
					}
				],
				"responses": {
					"200": {
						"description": "Successfully retrieved list of dogs",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "./dogs.json#/definitions/dogs"
								}
							}
						}
					}
				}
			},
			"post": {
				"tags": ["Dogs"],
				"summary": "Add a new dog to the database",
				"description": "Adds a new dog to the database using information provided",
				"requestBody": {
					"description": "Data to register a new dog",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "./dogs.json#/definitions/dog"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Dog was successfully added to the database",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "./dogs.json#/definitions/dogCreated"
								}
							}
						}
					},
					"400": {
						"$ref": "#/components/responses/validationError"
					},
					"401": {
						"$ref": "#/components/responses/authenticationError"
					},
					"403": {
						"description": "Only administrators may add dogs"
					}
				}
			}
		},
		"/dogs/{id}": {
			"description": "A single dog record",
			"parameters": [{ "$ref": "#/components/parameters/dogId" }],
			"get": {
				"tags": ["Dogs"],
				"summary": "Get a single dog entry",
				"security": [],
				"responses": {
					"200": {
						"description": "Dog entry successfully retrieved",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "./dogs.json#/definitions/dogData"
								}
							}
						}
					},
					"404": {
						"$ref": "#/components/responses/notFoundError"
					}
				}
			},
			"put": {
				"tags": ["Dogs"],
				"summary": "Update a dog entry",
				"description": "Updates a dog's entry by dog ID and data provided",
				"requestBody": {
					"description": "Data to update a dog",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "./dogs.json#/definitions/dog"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Dog was successfully updated",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "./dogs.json#/definitions/dogUpdated"
								}
							}
						}
					},
					"400": { "$ref": "#/components/responses/validationError" },
					"401": {
						"$ref": "#/components/responses/authenticationError"
					},
					"403": {
						"description": "Only administrators may alter dog entries"
					},
					"404": { "$ref": "#/components/responses/notFoundError" }
				}
			},
			"delete": {
				"tags": ["Dogs"],
				"summary": "Delete a dog from the database",
				"description": "Deletes a dog from the database by dog ID",
				"responses": {
					"200": {
						"description": "Dog was successfully deleted",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "./dogs.json#/definitions/dogDeleted"
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/authenticationError"
					},
					"403": {
						"description": "Only administrators may delete dog entries"
					},
					"404": { "$ref": "#/components/responses/notFoundError" }
				}
			}
		},
		"/dogs/{id}/breed": {
			"description": "A single dog's breed",
			"parameters": [{ "$ref": "#/components/parameters/dogId" }],
			"get": {
				"tags": ["Dogs"],
				"summary": "Get a dog's breed entry",
				"security": [],
				"responses": {
					"200": {
						"description": "Successflly retireved dog breed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "./dogs.json#/definitions/dogBreedData"
								}
							}
						}
					},
					"404": { "$ref": "#/components/responses/notFoundError" }
				}
			},
			"post": {
				"tags": ["Dogs"],
				"summary": "Assign a breed to the dog",
				"description": "Assigns a breed to a dog using the provided breed ID",
				"requestBody": {
					"description": "Dog ID and breed ID to link",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "./dogs.json#/definitions/dogBreed"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Successfully added dog breed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "./dogs.json#/definitions/dogBreedCreated"
								}
							}
						}
					},
					"400": { "$ref": "#/components/responses/validationError" },
					"401": {
						"$ref": "#/components/responses/authenticationError"
					},
					"403": {
						"description": "Only administrators may assign breeds"
					},
					"404": { "$ref": "#/components/responses/notFoundError" }
				}
			},
			"put": {
				"tags": ["Dogs"],
				"summary": "Update breed assigned to a dog",
				"description": "Updates a given dog's breed using provided breed ID",
				"requestBody": {
					"description": "Dog ID and breed ID to update",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "./dogs.json#/definitions/dogBreed"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Successfully updated dog breed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "./dogs.json#/definitions/dogBreedUpdated"
								}
							}
						}
					},
					"400": { "$ref": "#/components/responses/validationError" },
					"401": {
						"$ref": "#/components/responses/authenticationError"
					},
					"403": {
						"description": "Only administrators may assign breeds"
					},
					"404": { "$ref": "#/components/responses/notFoundError" }
				}
			},
			"delete": {
				"tags": ["Dogs"],
				"summary": "Delete a dog's breed assignment",
				"description": "Deletes a dog's breed entry by the dog's ID",
				"responses": {
					"200": {
						"description": "Dog breed successfully deleted",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "./dogs.json#/definitions/dogBreedDeleted"
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/authenticationError"
					},
					"403": {
						"description": "Only administrators may assign breeds"
					},
					"404": { "$ref": "#/components/responses/notFoundError" }
				}
			}
		},
		"/dogs/{id}/location": {
			"description": "Location of a dog",
			"parameters": [{ "$ref": "#/components/parameters/dogId" }],
			"get": {
				"tags": ["Dogs"],
				"summary": "Get the location a dog is associated with",
				"security": [],
				"responses": {
					"200": {
						"description": "Successfully retrieved dog location",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "./dogs.json#/definitions/dogLocationData"
								}
							}
						}
					},
					"404": { "$ref": "#/components/responses/notFoundError" }
				}
			},
			"post": {
				"tags": ["Dogs"],
				"summary": "Assigns a location to a dog",
				"description": "Assigns a location to a dog by taking a location ID",
				"requestBody": {
					"description": "Dog ID and location ID to link",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "./dogs.json#/definitions/dogLocation"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Successfully assigned location to a dog",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "./dogs.json#/definitions/dogLocationCreated"
								}
							}
						}
					},
					"400": { "$ref": "#/components/responses/validationError" },
					"401": {
						"$ref": "#/components/responses/authenticationError"
					},
					"403": {
						"description": "Only administrators may alter dogs"
					},
					"404": { "$ref": "#/components/responses/notFoundError" }
				}
			},
			"put": {
				"tags": ["Dogs"],
				"summary": "Assigns a new location to a dog",
				"description": "Assigns a given dog's location based on a location ID",
				"requestBody": {
					"description": "Dog ID and location ID to update",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "./dogs.json#/definitions/dogLocation"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Successfully updated dog location",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "./dogs.json#/definitions/dogLocationUpdated"
								}
							}
						}
					},
					"400": { "$ref": "#/components/responses/validationError" },
					"401": {
						"$ref": "#/components/responses/authenticationError"
					},
					"403": {
						"description": "Only administrators may alter dogs"
					},
					"404": { "$ref": "#/components/responses/notFoundError" }
				}
			},
			"delete": {
				"tags": ["Dogs"],
				"summary": "Delete a dog's location",
				"description": "Deletes a dog's location by dog ID",
				"responses": {
					"201": {
						"description": "Successfully deleted dog location",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "./dogs.json#/definitions/dogLocationDeleted"
								}
							}
						}
					},
					"400": { "$ref": "#/components/responses/validationError" },
					"401": {
						"$ref": "#/components/responses/authenticationError"
					},
					"403": {
						"description": "Only administrators may alter dogs"
					},
					"404": { "$ref": "#/components/responses/notFoundError" }
				}
			}
		},
		"/dogs/{id}/favourites": {
			"description": "A list of a dog's favourites",
			"parameters": [{ "$ref": "#/components/parameters/dogId" }],
			"get": {
				"tags": ["Dogs"],
				"description": "Get a list of favourites associated with the dog",
				"security": [],
				"responses": {
					"200": {
						"description": "Successfully retrieved favourites list",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "./dogs.json#/definitions/dogFavourites"
								}
							}
						}
					},
					"404": { "$ref": "#/components/responses/notFoundError" }
				}
			}
		},
...
		"/locations/{id}/dogs": {
			"description": "Dogs at a given location",
			"parameters": [{ "$ref": "#/components/parameters/locationId" }],
			"get": {
				"summary": "Get a list of dogs",
				"description": "Gets a list of dogs at a given location by location ID",
				"parameters": [
					{
						"name": "query",
						"description": "A search query",
						"in": "query",
						"required": false,
						"schema": { "type": "string" },
						"example": "Rex"
					},
					{
						"name": "page",
						"description": "Page number to browse",
						"in": "query",
						"required": false,
						"schema": { "type": "integer" },
						"example": 1
					},
					{
						"name": "limit",
						"description": "Limit of entries per page",
						"in": "query",
						"required": false,
						"schema": { "type": "integer" },
						"example": 5
					},
					{
						"name": "order",
						"description": "Item to order the results by",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string",
							"enum": ["id", "name", "description", "age", "gender"]
						},
						"example": "age"
					},
					{
						"name": "direction",
						"description": "Direction to sort in (ascending or descending)",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string",
							"enum": ["asc", "desc"]
						},
						"example": "asc"
					}
				],
				"security": [],
				"tags": ["Locations"],
				"responses": {
					"200": {
						"description": "Successfully retrieved list of dogs",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "./dogs.json#/definitions/dogs"
								}
							}
						}
					},
					"404": { "$ref": "#/components/responses/notFoundError" }
				}
			}
		},
...
	}
}